+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Avoidable overflow error when using NumpyPackedFSLDataset.
 - Fixed issue with NumpyFSLDatasetMixture + SourceMixtureDataset where not all instances would have the same sequence length.
 - Attention backend will no longer default to flash in non-CUDA environments.
+- Handle nan losses/gnorms in the `SkipStepOptimizer`.
 
 ### Changed